+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - File sharings ([#42](https://github.com/sohosai/sos21-backend/pull/42))
 
 ### Changed
+
+- Adjust file usage quota ([#43](https://github.com/sohosai/sos21-backend/pull/43))
+
 ### Deprecated
 ### Removed
 ### Fixed